requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
httpx==0.28.1
Jinja2==3.1.5
python-multipart==0.0.20
sqlalchemy==2.0.37
asyncpg==0.30.0
APScheduler==3.11.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4